parse_parameter: Continue parsing after 'mtdparts'.
The parse_paramter loop iterates over a parameter.txt file's contents
and parses that line-by-line. However, the current implementation
silently stops after parsing 'mtdparts'. This is problematic if
a 'uuid' token is found in one of the following lines.

This patch addresses the problem, by simply continue to iterate
over the next lines after 'mtdparts'.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>